rename memory.c and parser.c to be less generic; eliminate hash.c
