+++ Makefile.common
@@ -1952,9 +1952,8 @@ ifeq ($(HAVE_NETWORKING), 1)
 
       OBJ += cheevos/cheevos.o \
              cheevos/badges.o \
-             cheevos/memory.o \
-             cheevos/parser.o \
-             cheevos/hash.o \
+             cheevos/cheevos_memory.o \
+             cheevos/cheevos_parser.o \
              $(LIBRETRO_COMM_DIR)/formats/cdfs/cdfs.o \
              deps/rcheevos/src/rcheevos/alloc.o \
              deps/rcheevos/src/rcheevos/compat.o \
