+++ cheevos/cheevos_parser.c
@@ -1,6 +1,5 @@
-#include "parser.h"
+#include "cheevos_parser.h"
 
-#include "hash.h"
 #include "util.h"
 
 #include <encodings/utf.h>
@@ -33,6 +32,17 @@
 /*****************************************************************************
 Gets a value in a JSON
 *****************************************************************************/
+
+static uint32_t rcheevos_djb2(const char* str, size_t length)
+{
+   const unsigned char* aux = (const unsigned char*)str;
+   uint32_t            hash = 5381;
+
+   while (length--)
+      hash = (hash << 5) + hash + *aux++;
+
+   return hash;
+}
 
 typedef struct
 {
